requirements.txt
pydantic==1.9.0
fastapi==0.74.0
uvicorn==0.17.5
python-multipart
jinja2
websockets==10.1
starlette==0.17.1
pyserial
loguru
supervisor
pyyaml
redis
sqlalchemy==1.4.31
fire
psycopg2-binary==2.9.3
retrying
requests
pytz
numpy
pyjwt
passlib~=1.7.4
sqlalchemy-utils
google-cloud-texttospeech
pydub
crcmod
opencv-python
squareup
mutagen